fix: improve online users, fix bug for other url page not show users online
